chore: Move BloomMetadata to own file
